Fix bug in process picker, make 8-key mode more restrictive and apply to 4-keys
Windowfinder was incorrectly continuing to search for valid windows,
breaking if the process was not "LoveRitmo". Fixed by exiting the loop
after a match is found (expected behaviour)

Changed eight-key mode to affect arrows (4-key mode) as well. If
eight-key mode and numlock is enabled, it will send the proper numpad
key (i.e numpad2 instead of arrow down). This is not necessary, but it
more closely represents how a human would play in 8-key mode (I guess)